cleaned up some tests
